+++ CHANGELOG.md
@@ -33,6 +33,10 @@ Line wrap the file at 100 chars.                                              Th
 - Old `mullvad log set-level` command has been renamed to `mullvad log set-rust-log`.
 - Update `gotatun` to 0.8.1.
 
+#### Linux
+- Make all timestamps embedded in `.deb` and `.rpm` packages deterministic by deriving them from
+  the git commit being built instead of the time of the build. Required for reproducible builds.
+
 #### Windows
 - Update `wireguard-nt` to version 1.1. This retires the Mullvad fork at
   <https://github.com/mullvad/wireguard-nt>.